backend/requirements.txt
mh_z19
paho-mqtt

